requirements.txt
Django==5.2.6
django-ninja==1.4.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0
django-cors-headers==4.8.0
requests==2.32.5